Issue 36: Add filter threshold
Add filter threshold. Courses with an activity count over the
threshold will default to the first activity type. Add a threshold
setting that defaults to 40. Add a warning that a large number of
activities will cause the page to take a long time to display. Append
number of activities to the all activities option.